Create drawpitch.py
Not an impressive bit of code, but if you are doing rugby analysis. This can be a great tool for visualizing. It was really tedious to make this rugby pitch so hopefully someone else can take this and make some use out of it. I laid a heatmap over it to track certain players